Modified AttayMockLoad.py:
1.  Force dependent directories to be in same directory as the
ArrayMockLoad.py file.  This means that the ArrayMockLoad.py file should
be callable from pywbemcli when it is in another directory.

2. Modified the structure to use the pywbem mock script interface
defined in pywbem 1.0 and available in pywbemcli 0.8.0. This the
interface from using globals for CONN, SERVER, and VERBOSE to using a
function interface and putting the code into that method.  The method
interface is:

setup(conn, server, verbose)

3. Modifies some code to simplify it.

Makes some simple modifications to BuildLeafClasses.py

Add a .gitignore

For some reason it was an empty directory. The directory was removed.